реализация поиска, стили активного меню, переход по якорю
