+++ src/components/Main/Promo/Promo.css
@@ -56,6 +56,9 @@
     padding: 10px;
     transition: opacity 0.7s linear;
     grid-area: button;
+    box-sizing: border-box;
+    text-decoration: none;
+    text-align: center;
 }
 
 .promo__button:hover {
