+++ src/components/Movies/Movies.js
@@ -11,15 +11,15 @@ function Movies({ setMovies,
       // handleChange,
       isValid,
       setIsValid,
-      // errors,
+      errors,
       // setErrors,
       setIsSearch,
       isSearch,
       setSearchPhrase,
       searchPhrase,
       searchWord,
-      // setSearchMovie,
-      // searchMovie,
+      setSearchMovie,
+      searchMovie,
       // getPhilms,
       getSavedMovies,
       setSearchMessage,
@@ -36,7 +36,7 @@ function Movies({ setMovies,
       //setCurrentUser,
       //currentUser,
       arrayLikeMovieId,
-      // handleChangeSearchPhrase,
+      handleChangeSearchPhrase,
       handlerCheckBox,
       setHeartRed,
       isHeartRed,
@@ -45,62 +45,42 @@ function Movies({ setMovies,
       setButtonMore,
       buttonMore,
       indexByWidth,
-      // movieSearch
+      movieSearch,
+      handleSearchSubmit,
+      searchSubmitAndCheck
 
 }) {
 
-      // const [searchPhrase, setSearchPhrase] = useState(null);
-      const [values, setValues] = useState({});
-      const [errors, setErrors] = useState('');
-      const [searchMovie, setSearchMovie] = useState([]);
 
-      useEffect(() => {
-            setSearchPhrase(localStorage.getItem('searchPhrase'));
-      }, [])
 
-      useEffect(() => {
-            if (movies) {
-                  // setSearchPhrase(localStorage.getItem('searchPhrase'));
-                  movieSearch()
-                  setSearchMessage('');
-            };
-      }, [movies]);
+      // const [searchMovie, setSearchMovie] = useState(JSON.parse(localStorage.getItem('searchMovie')) || []);
+      
+      // useEffect(() => {
+      //       setSearchMovie(JSON.parse(localStorage.getItem('searchMovie')));
+      // }, [setSearchMovie])
 
-      function movieSearch() {
-            setIsLoading(true);
-            let newMovie = [];
-            if (movies && searchPhrase) {
-                  movies.map((movie) => {
-                        let nameRU = movie.nameRU.toLowerCase();
-                        if (isCheckBox && nameRU.includes(searchPhrase.toLowerCase())) {
-                              newMovie.push(movie);
-                        } else if (!isCheckBox && nameRU.includes(searchPhrase.toLowerCase())
-                              && movie.duration >= 40) {
-                              newMovie.push(movie);
-                        }
-                  })
-                  if (newMovie.length === 0) {
-                        console.log('ошибка');
-                        setSearchMessage('Ничего не найдено!')
-                  }
-                  localStorage.setItem('searchMovies', JSON.stringify(newMovie));
-                  setSearchMovie(newMovie);
-            }
-            //   setTimeout(showLoader, 1500);
-            setIsLoading(false);
-      }
+
+      // useEffect(() => {
+      //       if (movies) {
+      //             setSearchPhrase(localStorage.getItem('searchPhrase'));
+      //             movieSearch()
+      //             setSearchMessage('');
+      //       };
+      // }, [movies]);
+
+      
 
 
 
-      const handleChangeSearchPhrase = (event) => {
-            const target = event.target;
-            const name = target.name;
-            const value = target.value;
-            setValues({ ...values, [name]: value });
-            setErrors(target.validationMessage);
-            setIsValid(target.closest("form").checkValidity());
-            setSearchPhrase(event.target.value);
-      }
+      // const handleChangeSearchPhrase = (event) => {
+      //       const target = event.target;
+      //       const name = target.name;
+      //       const value = target.value;
+      //       setValues({ ...values, [name]: value });
+      //       setErrors(target.validationMessage);
+      //       setIsValid(target.closest("form").checkValidity());
+      //       setSearchPhrase(event.target.value);
+      // }
 
 
       return (
@@ -140,6 +120,8 @@ function Movies({ setMovies,
                               movieSearch={movieSearch}
                               setIsLoading={setIsLoading}
                               isLoading={isLoading}
+                              handleSearchSubmit={handleSearchSubmit}
+                              searchSubmitAndCheck={searchSubmitAndCheck}
                         />
                         {isLoading ?
                               <Preloader />
