+++ src/components/Movies/MoviesCardList/MovieCardList.js
@@ -38,7 +38,7 @@ function MovieCardList({
     
     const [inputMovie, setInputMovie] = useState([]);
 
-    console.log('savedMovies in MovieCardList-->', savedMovies)
+    console.log(searchMovie);
 
     function handleMore() {
         if(window.innerWidth >= 1158) {
@@ -51,13 +51,6 @@ function MovieCardList({
             setIndex(i => i + 1);
         }
     } 
-
-    // useEffect(() => {
-    //     if(!savedMoviesPage) {
-    //     setSearchPhrase('ла');
-    //     console.log('useeffect movicardlist searchPhrase ->', searchPhrase);
-    //     }
-    // }, [])
 
     useEffect(() => {
         if (!savedMoviesPage) {
