+++ src/components/Navigation/Navigation.css
@@ -21,6 +21,10 @@
     transition: opacity 0.7s linear;
 }
 
+.button__active {
+    border: 1px solid;
+}
+
 .button:hover {
     opacity: 0.5;
     cursor: pointer;
